approval_reconciliation api endpoint
adding an endpoint for future admin dashboard changes to allow admins to compare and contrast similar incident ids.  See trello card details 
As an Admin-User I want to see the highest occurring and agreed upon location in the Admin Dashboard for an incident we've requested more information on. Determine how to handle consensus on suggestions for location information on incidents we request more information on via form.