Update main.py
Builds out endpoints to see admin dashboard different data on incidents ids with differing state, locations, and time.  Driven by the new get_root_twelve_majority function which attempts to satisfy the trello user story 

As an Admin-User I want to see the highest occurring and agreed upon location in the Admin Dashboard for an incident we've requested more information on. Determine how to handle consensus on suggestions for location information on incidents we request more information on via form.